airflow/requirements.txt
absl-py==1.2.0
alembic==1.8.0
amqp==5.1.1
anyio==3.6.1
apache-airflow==2.3.3
apache-airflow-providers-celery==3.0.0
apache-airflow-providers-ftp==3.0.0
apache-airflow-providers-http==3.0.0
apache-airflow-providers-imap==3.0.0
apache-airflow-providers-sqlite==3.0.0
apispec==3.3.2
argcomplete==2.0.0
astunparse==1.6.3
attrs==20.3.0
Babel==2.10.3
billiard==3.6.4.0
blinker==1.4
cachelib==0.9.0
cachetools==5.2.0
cattrs==1.10.0
celery==5.2.7
certifi==2020.12.5
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clickclick==20.10.2
colorama==0.4.5
colorlog==4.8.0
commonmark==0.9.1
connexion==2.14.0
cron-descriptor==1.2.30
croniter==1.3.5
cryptography==36.0.2
cycler==0.11.0
Deprecated==1.2.13
dill==0.3.1.1
dnspython==2.2.1
docutils==0.19
email-validator==1.2.1
Flask==2.1.2
Flask-AppBuilder==4.1.2
Flask-Babel==2.0.0
Flask-Caching==2.0.0
Flask-JWT-Extended==4.4.2
Flask-Login==0.6.1
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
flatbuffers==1.12
flower==1.1.0
fonttools==4.34.4
gast==0.4.0
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20
greenlet==1.1.2
grpcio==1.47.0
gunicorn==20.1.0
h11==0.12.0
h5py==3.7.0
httpcore==0.15.0
httpx==0.23.0
humanize==4.2.3
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.6.1
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
kombu==5.2.4
lazy-object-proxy==1.7.1
libclang==14.0.1
linkify-it-py==2.0.0
lockfile==0.12.2
Mako==1.2.1
Markdown==3.3.7
markdown-it-py==2.1.0
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.5.2
mdit-py-plugins==0.3.0
mdurl==0.1.1
numpy==1.23.1
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
pathspec==0.9.0
pendulum==2.1.2
Pillow==9.2.0
pluggy==1.0.0
prison==0.2.1
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-daemon==2.3.0
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==6.1.2
pytz==2022.1
pytzdata==2020.1
PyYAML==6.0
requests==2.28.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==12.4.4
rsa==4.9
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
setproctitle==1.2.3
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.27
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.2
swagger-ui-bundle==0.0.9
tabulate==0.8.10
tenacity==8.0.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==3.1.0
tornado==6.2
typing-extensions==4.3.0
uc-micro-py==1.0.1
unicodecsv==0.14.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
WTForms==2.3.3
zipp==3.8.0
